+++ src/main/java/org/eclipse/lyo/oslc4j/trs/server/PagedTrsFactory.java
@@ -1,0 +1,6 @@
+package org.eclipse.lyo.oslc4j.trs.server;
+
+public interface PagedTrsFactory {
+    PagedTrs getPagedTrs();
+    ResourceEventHandler getResourceEventHandler();
+}