+++ src/main/java/org/eclipse/lyo/oslc4j/trs/server/service/TrackedResourceSetService.java
@@ -1,5 +1,5 @@
 /*-
- * Copyright (c) 2016-2017   KTH Royal Institute of Technology.
+ * Copyright (c) 2016-2019   KTH Royal Institute of Technology.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -12,7 +12,7 @@
  * Contributors:
  *
  * Omar Kacimi         -  Initial implementation
- * Andrew Berezovskyi  -  Lyo contribution updates
+ * Andrew Berezovskyi  -  Lyo contribution updates; refactoring.
  */
 
 package org.eclipse.lyo.oslc4j.trs.server.service;
@@ -25,19 +25,19 @@ import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
-import javax.ws.rs.core.UriInfo;
+import javax.ws.rs.core.UriBuilder;
 import org.eclipse.lyo.core.trs.Base;
 import org.eclipse.lyo.core.trs.ChangeLog;
 import org.eclipse.lyo.core.trs.Page;
 import org.eclipse.lyo.core.trs.TRSConstants;
 import org.eclipse.lyo.core.trs.TrackedResourceSet;
+import org.eclipse.lyo.oslc4j.core.OSLC4JUtils;
 import org.eclipse.lyo.oslc4j.core.annotation.OslcService;
 import org.eclipse.lyo.oslc4j.core.model.Error;
 import org.eclipse.lyo.oslc4j.core.model.OslcMediaType;
-import org.eclipse.lyo.oslc4j.trs.server.IChangeHistories;
+import org.eclipse.lyo.oslc4j.trs.server.PagedTrs;
 import org.eclipse.lyo.oslc4j.trs.server.TRSUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -49,27 +49,26 @@ import org.slf4j.LoggerFactory;
  * @version $version-stub$
  * @since 2.3.0
  */
-@Path("/trs")
+@Path(TrackedResourceSetService.RESOURCE_PATH)
 @OslcService(TRSConstants.TRS_NAMESPACE)
 public class TrackedResourceSetService {
     private static final Logger log     = LoggerFactory.getLogger(TrackedResourceSetService.class);
-    public static final String BASE_PATH = "base";
-    public static final String CHANGELOG_PATH = "changeLog";
+    private static final String BASE_PATH = "base";
+    private static final String CHANGELOG_PATH = "changeLog";
+    public static final String RESOURCE_PATH = "/trs";
 
     /**
      * The instance of the change histories class used by a trs service class implementing this
      * class. The instance returned is expected to be a singleton of a class implementing the
      * ChangeHistories class
-     *
-     * @return the instance of the class implementing the change histories class
      */
-    private IChangeHistories changeHistories;
+    private PagedTrs changeHistories;
 
     public TrackedResourceSetService() {
     }
 
     @Inject
-    public TrackedResourceSetService(IChangeHistories _changeHistories) {
+    public TrackedResourceSetService(PagedTrs _changeHistories) {
         changeHistories = _changeHistories;
     }
 
@@ -81,23 +80,34 @@ public class TrackedResourceSetService {
     @GET
     @Produces({OslcMediaType.TEXT_TURTLE, OslcMediaType.APPLICATION_RDF_XML,
             OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON})
-    public TrackedResourceSet getTrackedResourceSet(@Context UriInfo uriInfo) {
+    public TrackedResourceSet getTrackedResourceSet()
+            throws URISyntaxException {
         TrackedResourceSet result = new TrackedResourceSet();
 
-        result.setAbout(uriInfo.getRequestUri());
-        result.setBase(uriInfo.getAbsolutePathBuilder().path(BASE_PATH).build());
+        result.setAbout(uriBuilder().build());
+        result.setBase(uriBuilder().path(BASE_PATH).build());
 
-        ChangeLog changeLog = getChangeHistories().getChangeLog("1");
-        if (changeLog == null) {
-            changeLog = new ChangeLog();
-        }
-        try {
-            result.setChangeLog(changeLog);
-        } catch (URISyntaxException e) {
+        if(getPagedTrs().changelogPageCount() == 0) {
             // FIXME Andrew@2019-04-27: remove this exception from the signature
-            throw new IllegalArgumentException("Can't set the change log", e);
+            result.setChangeLog(new ChangeLog());
+        } else {
+            result.setChangeLog(getPagedTrs().getChangeLogLast());
         }
+
         return result;
+    }
+
+    /**
+     * manage http calls for the first page of the base. The call is redirected to the handler of
+     * http calls for a specific base page as a call for the page 1 of the base
+     *
+     * @return the first page of the base
+     */
+    @GET
+    @Path(BASE_PATH)
+    public Response getBase() {
+        final URI newURI = uriBuilder().path(BASE_PATH).path("1").build();
+        return Response.seeOther(newURI).build();
     }
 
     /**
@@ -110,28 +120,50 @@ public class TrackedResourceSetService {
     @Path(BASE_PATH + "/{page}")
     @Produces({OslcMediaType.TEXT_TURTLE, OslcMediaType.APPLICATION_RDF_XML,
                       OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON})
-    public Response getBasePage(@PathParam("page") String pageNo) {
-        Base base = getChangeHistories().getBaseResource(pageNo);
+    public Response getBasePage(@PathParam("page") int pageNo) {
+        Base base = getPagedTrs().getBaseResource(pageNo);
         if (base == null) {
             final Error entity = new Error();
             entity.setMessage("Wrong TRS Base page URI");
             entity.setStatusCode(String.valueOf(Status.NOT_FOUND.getStatusCode()));
             return Response.status(Status.NOT_FOUND).entity(entity).build();
         }
-        // Due to OSLC4J limitation, not Base but NextPage will be returned.
-        // See org.eclipse.lyo.rio.trs.resources.BaseResource.getBasePage(Long)
-        // FIXME Andrew@2019-04-27: do the math right
+
+        // Return the nextPage Page object, which describes the next base page in terms,
+        // of the current base page we are manipulating.  We do not directly 
+        // return the base object due to a limitation in OSLC4J.  Currently 
+        // OSLC4J requires that triples in the RDF graph with different subjects
+        // reference one another.  According to the 2.0 spec, the Page object
+        // already references the Base object so we will get the appropriate
+        // output if we return Page.  If we force a reference from Base to Page
+        // instead then we get a ldp:nextPage entry which does not conform to the
+        // TRS 2.0 specification.
+        // TODO Andrew@2019-05-01: fix it
         Page nextPage = base.getNextPage();
         if (nextPage == null) {
             throw new WebApplicationException(Status.NOT_FOUND);
         }
         log.debug("TRS Base page contains {} members", base.getMembers().size());
-        return Response.ok(nextPage).header("Link", TRSUtil.linkHeaderValue(base)).build();
+        return Response.ok(base).header("Link", TRSUtil.linkHeaderValue(base)).build();
     }
 
-    protected IChangeHistories getChangeHistories() {
+    protected PagedTrs getPagedTrs() {
         return changeHistories;
     }
+
+    /**
+     * manage the calls for the change log and redirects to the handler of a specific page of the
+     * change log with the call to the first page
+     *
+     * @return the first page of the change lof
+     */
+    @GET
+    @Path(CHANGELOG_PATH)
+    public Response getChangeLog() {
+        final URI newURI = uriBuilder().path(CHANGELOG_PATH).path("1").build();
+        return Response.seeOther(newURI).build();
+    }
+
 
     /**
      * Returns the requested page of the change log
@@ -143,11 +175,10 @@ public class TrackedResourceSetService {
     @Path(CHANGELOG_PATH + "/{page}")
     @Produces({OslcMediaType.TEXT_TURTLE, OslcMediaType.APPLICATION_RDF_XML,
                       OslcMediaType.APPLICATION_XML, OslcMediaType.APPLICATION_JSON})
-    public Response getChangeLogPage(@PathParam("page") String page) {
+    public Response getChangeLogPage(@PathParam("page") int page) {
         log.trace("TRS Change Log page '{}' requested", page);
 
-        ChangeLog changeLog = null;
-        changeLog = getChangeHistories().getChangeLog(page);
+        ChangeLog changeLog = getPagedTrs().getChangeLog(page);
         if (changeLog == null) {
             final Error entity = new Error();
             entity.setMessage("Wrong TRS Change Log page URI");
@@ -158,29 +189,7 @@ public class TrackedResourceSetService {
         return Response.ok(changeLog).build();
     }
 
-    /**
-     * manage the calls for the change log and redirects to the handler of a specific page of the
-     * change log with the call to the first page
-     *
-     * @return the first page of the change lof
-     */
-    @GET
-    @Path(CHANGELOG_PATH)
-    public Response getChangeLog(@Context UriInfo uriInfo) {
-        final URI newURI = uriInfo.getAbsolutePathBuilder().path("1").build();
-        return Response.seeOther(newURI).build();
-    }
-
-    /**
-     * manage http calls for the first page of the base. The call is redirected to the handler of
-     * http calls for a specific base page as a call for the page 1 of the base
-     *
-     * @return the first page of the base
-     */
-    @GET
-    @Path(BASE_PATH)
-    public Response getBase(@Context UriInfo uriInfo) {
-        final URI newURI = uriInfo.getAbsolutePathBuilder().path("1").build();
-        return Response.seeOther(newURI).build();
+    private static UriBuilder uriBuilder() {
+        return UriBuilder.fromUri(OSLC4JUtils.getServletURI()).path(RESOURCE_PATH);
     }
 }